+++ src/game/tcm.cpp
@@ -136,7 +136,7 @@ int TCM::sendi3cmd()
 
   for (const auto& cmd : cmds) {
     std::string msg = "i3-ipc";
-    uint32_t len = cmd.size();
+    uint32_t len = static_cast<uint32_t>(cmd.size());
     uint32_t type = 0;
 
     msg.append(reinterpret_cast<const char*>(&len), 4);
